Complete modulisation of the javascript
